Finished frontcontroller (should work properly now)

+++ POS Web Interface/src/controller/handler/DeletePersonHandler.java
@@ -6,24 +6,26 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import annotation.RequestMapping;
-import db.WebshopFacade;
 import domain.person.Person;
 
+@RequestMapping(action="deleteperson_start")
 @RequestMapping(action="deleteperson_complete")
 public class DeletePersonHandler extends Handler {
 
-	public DeletePersonHandler(WebshopFacade webshopFacade) {
-		super(webshopFacade);
-	}
-
 	@Override
-	public String handleRequest(HttpServletRequest request,
+	public String handleRequest(String action, HttpServletRequest request,
 			HttpServletResponse response) {
+		if(action.equals("deleteperson_start")) {
+			request.setAttribute("person",
+					webshopFacade.getPerson(request.getParameter("mail")));
+			return "deleteperson.jsp";
+		}
+		
 		webshopFacade.deletePerson(request.getParameter("mail"));
 		
 		List<Person> persons = webshopFacade.getPersons();
 		request.setAttribute("persons", persons);
-		return "personoverview.jsp";
+		return "personoverview";
 		
 		
 	}
